Rework the VM for inheritance, clarity and flexibility.
Just providing methods into the constructor is a very obtuse way of defining a VM. All the logic sits in different places and only Dagger really knows how the thing is configured. This makes discovery difficult, especially when new to the pattern.

I've changed it so that VM implementers only need to define the bare minimum of functions via overrides.