MVI Refactoring (#177)
* First round of refactoring

- Make VM implementations dumb
- Move mechanisms into base classes
- Require injection of Reducer/ActionTransformer

Still needs a better place for the Event->Action transformation.

Unit testing of ViewModel no longer makes much sense; ActionTranformer and Reducers are the only functionality specific parts to be tested.

Unit test of HFVM is broken.

* Further improvements

- Operation Failure is now a data class
- Threading is now explicitly controlled by the part requiring it (don't just shovel things to computation)

* - Use base interfaces for event types
- Add nicer logger

* Don't emit the same thing twice from user fetch

* Remove logging parameter from MVI events

* Combine log tag (only required because Java modules)

* Move top-level bits to a single file

* Move into separate directory

* Use functions, not classes

* More API neatening

* More provider to Mvi package

* More typing

* Get rid of Chuck and reorder type parameters

* Move to arch package

* Restructure arch packages

* Restructure arch packages

* Core arch

* Log Fatal error as an error

* Formatting

* Move mvi vm to separate package

* Remove Renderer from MVI

* Rename to dispatcher

* Rename to dispatcher

* Use destructuring in fragment

* Convert basefragment to kotlin

* Use Flow and make VM a Fragment type parameter

* Separate dispatcher from store

* Naming

* Privatize the fields

* Absolutely no idea what has changed. It's been months.

* Rework the VM for inheritance, clarity and flexibility.

Just providing methods into the constructor is a very obtuse way of defining a VM. All the logic sits in different places and only Dagger really knows how the thing is configured. This makes discovery difficult, especially when new to the pattern.

I've changed it so that VM implementers only need to define the bare minimum of functions via overrides.

* Rework the VM for inheritance, clarity and flexibility.

Just providing methods into the constructor is a very obtuse way of defining a VM. All the logic sits in different places and only Dagger really knows how the thing is configured. This makes discovery difficult, especially when new to the pattern.

I've changed it so that VM implementers only need to define the bare minimum of functions via overrides.

* Fix up the test, introduce some boilerplate to impls classes.

Now need to call: bind() from the child classes.

Using inheritance, it seems unavoidable, because if the binding is triggered by the parent class instance in its constructor, then the child class will not have been initialized and its dependencies will be null.

Perhaps something could be done using delegates?

* Tighten visibility of ViewModel abstract methods

* Travis is fun

* Moderize the travis file

* Fix naming weirdness

* Don't run device tests in Travis

* Disable sonarqube for now, this project got removed.